requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Firebase Auth
firebase-admin==6.4.0

# Anthropic Claude
anthropic==0.18.1

# Image processing
Pillow==10.2.0
pdf2image==1.17.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
